+++ messaging/coap/oc_coap.h
@@ -39,8 +39,8 @@ struct oc_separate_response_s
 struct oc_response_buffer_s
 {
   uint8_t *buffer;
-  uint16_t buffer_size;
-  uint16_t response_length;
+  size_t buffer_size;
+  size_t response_length;
   int code;
   oc_content_format_t content_format;
 };
